refactor models

+++ src/pkg/models/estimate/estimate.go
@@ -1,6 +1,6 @@
 package estimate
 
 type Estimate struct {
-	MilestoneID	uint	`json:"milestone_id" bson:"milestone_id"`
+	MilestoneID	uint64	`json:"milestone_id" bson:"milestone_id"`
 	EstimateURL	string	`json:"estimate_url" bson:"estimate_url"`
 }
